Don't set permissions on pre-existing parent dir (#51)
When a non-default file sink mode is configured, the parent directory's
permissions are also modified to match the configured mode. If the
desired mode does not include execute permissions, the permission
change effectively removes the ability for a non-root user to enter
into the directory.

This change makes it so that only the created file's permissions are
modified. The parent directory's permissions are unchanged if the
directory already exists.