fix (filters/encrypt): Properly support Taggable filter operations
The initial implementation of Taggable required the
all impl. of Taggable to return the filter operation for each
tagged field and ignored the node's configured and overridden
filter operations.  This fixes that issue.  Going fwd the impl. of
Taggable "may" optionally return a filter op, and if none is returned
the node's filter operations and/or overrides are used.