Support tagged fields and taggable (#60)
* fix (filters/encrypt): make a copy of the event data before modifying it

since the node will be modifying the event data (aka redact/encrypt), we
need our own copy, otherwise we could be changing the event across other
pipelines and nodes and creating a host of problems and race conditions.

* fix (filters/encrypt): skip non-exported fields

When filtering fields, skip the non-exported fields which will
significantly improve prerformance when dealing with things
like protobufs which can have MANY unexported fields with
MANY values.

* feat (filters/encrypt): add support for filtering structpb.Value

Add support for tagging maps which are google.protobuf.Struct and use
structpb.Value for their values.

* feat (filters/encrypt): support structs with tagged fields and taggable interface

Support structs that have both tagged fields and have also
implemented the Taggable interface for its fields which
are maps

* fix (filters/encrypt): fix check for top level Taggable interface

We need to check if the top level payload interface{} for an
event is Taggable before we strip off the interface to the
underlying value.

* test (filters/encrypt): Add tests for protobuf payloads

* feat (filters/encrypt): Support filtering proto string/btye wrappers

* fix (filters/encrypt): Properly support Taggable filter operations

The initial implementation of Taggable required the
all impl. of Taggable to return the filter operation for each
tagged field and ignored the node's configured and overridden
filter operations.  This fixes that issue.  Going fwd the impl. of
Taggable "may" optionally return a filter op, and if none is returned
the node's filter operations and/or overrides are used.