feat (encrypt): Add filtering of maps which don't impl. Taggable
Currently, map fields are only filtered if a Taggable interface is
implemented which includes the map. This is efficient, but could result
in an inadvertent leak of sensitive and/or secret data.  This commit
changes that behavior by tracking all maps which are encountered while
processing the event and before returning, it filters all the
remaining unfiltered maps.