feat (encrypt): Setting filters to no operation.
Add support to set a data classification's filter operation to
encrypt.NoOperation which was a use-case that hadn't been supported.

Also, if all filter operations == encrypt.NoOperation, then short
circuit the processing of the event and return the original event.
This will save a ton of unnecessary processing when no filtering is
required.