Events: filter map fields which are not tagged. (#63)
* feat (encrypt): Add filtering of maps which don't impl. Taggable

Currently, map fields are only filtered if a Taggable interface is
implemented which includes the map. This is efficient, but could result
in an inadvertent leak of sensitive and/or secret data.  This commit
changes that behavior by tracking all maps which are encountered while
processing the event and before returning, it filters all the
remaining unfiltered maps.

* fix (encrypt): Add missing options in calls to filterValue(...)

Both filterTaggable(...) and filterSlice(...) were not passing along
all the appropriate options when calling filterValue(...)
In particular, the wrapper options were not being passed along, which
meant that when data was encrypted or hmac-sha256'd it could have
used the incorrect optional wrapper.

* refactor (encrypt): Consistent use of TestMapField

Replace literal with a consistent use of the const
TestMapField.

* feat (encrypt): Filter maps/fields in Taggable which are not classified

While processing, via Filter.filterTaggable(...), the fields returned
from Taggable.Tags() keep "track" of each map and field which
has been filtered.

After filtering all the "tagged" fields, then filter all fields from
maps which were not classified.

This basically means if you reference a map within a json pointer
in an implementation of Taggable, we will make sure there isn't
any inadvertent data leaked within the fields you didn't "tag" via
a json pointer.

* refactor (encrypt): Make trackedMaps concurrent safe

* refact (encrypt): track maps found during processing

As we're processing an event, we'll track all the maps found
(encountered) and then process all unfiltered map fields
before returning the filtered event.

* docs (encrypt): Updated to reflect how Taggable maps work.