Use sync.Map for graph roots map (#81)
I ran into a race condition related to the `graph.roots` map, e.g.,

https://app.circleci.com/pipelines/github/hashicorp/vault/50560/workflows/96d003c2-185e-4504-a4c0-fab481ed6282/jobs/594936

happens when registering and removing a pipeline (which accesses
a `graph`'s `roots` map) when another goroutine is iterating through
the map in, for example, `graph.process()`. The pipeline registering
and removing are protected by the broker's lock, but the
`graph.process()` is *not* protected by the lock (since it does not
have access to it).

Adding another lock for `graph.roots` might be messy and lead to deadlock
possibilities, so I thought the easiest way to resolve this was
to switch to a `sync.Map`.

I confirmed that this change fixes the above race condition in
Vault's tests, and added a test here that reproduces the race
condition when run with `-race`.